LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/geolocation-API/enabled-by-permission-policy-attribute-redirect-on-load.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Feature policy 'Geolocation' check failed for element with origin 'https://127.0.0.1:9443' and allow attribute ';geolocation'.


PASS Permissions-Policy allow="geolocation" allows same-origin relocation
FAIL Permissions-Policy allow="geolocation" allows cross-origin relocation assert_true: geolocation expected true got false
